Factor out subprocess submodule
