+++ src/compiler/subprocess.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int run_subprocess(const char *name, const char **args);
